Update project dependencies, security settings and migration event logs
The commit involves updates in the pom.xml file, upgrading the spring-boot-starter-parent from 2.7.3 to 3.3.0, and java.version from 18 to 21. Also, the MySQL connector dependency was changed, and its version was pinned to 8.4.0. Moreover, the password was removed from both the pom.xml file and application.properties for better security. Lastly, changes in App.java and FlywayCallback.java were made to adjust the output log messages during the database migration events.