Update email system and Spring version
This commit updates the email references from placeholder emails to a valid email address and changes some specifics in the application.properties file to match. Also, the app Spring and Java versions were updated in the pom.xml file, and equivalent updates were implemented in the MailService.java file. Message variables in the templates were changed to English.