Refactor code and update dependencies
This commit includes multiple changes, mainly code cleanup and dependency update. Spaces are added to the HTML code for better readability and consistency. Code in many java files is refactored, changing the way of dependency injection to constructor-based. Also, the version of the Spring Boot dependency and java version in pom.xml are updated.