Update codebase for package dependencies and better Java conventions
This commit includes the upgrade of various package versions in pom.xml, including spring-boot-starter-parent and java.version for maintaining up-to-date dependencies, adding @Repository annotation in UserEntityRepository for better recognition, and modification of property injection to constructor injection in UserController for better Java conventions. It also improves security configuration and updates HTML templates for better user experience.