Update to Spring Boot 3.3.0, refactor JpaRepository imports, and apply code improvements
This commit updates the projects to use Spring Boot 3.3.0 from 2.7.3 and upgrades Java version to 21. Refactorings were implemented to use Jakarta Persistence API instead of the previous javax.persistence for the Customer entity. Code improvements include the use of constructor injection for better testability, removal of unused imports, corrections in comments translations, and jQuery version update. The Maven Central repository was added to the POM files.