Update Spring Boot version and OAuth2 security configurations
This commit is to update the Spring Boot version in both app1 and app2 pom.xml files to 3.3.0 and change java version to 21. OAuth2 security configurations in application.properties files of app1 and app2 have been modified to dynamically use clientId and clientSecret. Also, added .env in gitignore file and enhanced readability of SecurityConfig.java files.