Update Java packages and improve code quality
The commit updates the Java packages from javax.persistence to jakarta.persistence and enhances the quality of the code in several ways. Specifically, it updates the version in the Spring Boot Parent from 2.7.3 to 3.3.0 and the Java version from 18 to 21. It also adds NonNull annotations to the SecurityAdvice class for better null safety, switches the comment language from Spanish to English for better understanding, and refactors the autowiring in the CustomerController class.