Update spring-webflux module and refactor services and controllers
This commit updates the spring-boot-starter-parent version to 3.3.0 and the java version to 21 in the spring-webflux module. The code for services and controllers was refactored to utilise constructor injection instead of field injection, enhancing testability and following best practices. Formatting issues have also been corrected in the process for better code readability.